# libs/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trader API",
    "summary": "Исследовательская платформа по рыночным данным",
    "description": "REST API платформы. Документация: `/docs` (Swagger UI), `/redoc`, схема — `/openapi.json`.\n\n**WebSocket** `/ws/jobs/{job_id}` (в OpenAPI не описывается): шлёт JSON в формате `JobOut` при каждом изменении задачи и закрывается по её завершении; код закрытия 4404 — задачи нет.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/chart-profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Профили графика",
        "description": "С `instrument_id` — глобальные и профили этого инструмента; без него — все. Последние применённые первыми.",
        "operationId": "listChartProfiles",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по инструменту",
              "title": "Instrument Id"
            },
            "description": "Фильтр по инструменту"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileOut"
                  },
                  "title": "Response Listchartprofiles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Сохранить профиль",
        "operationId": "createChartProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          },
          "409": {
            "description": "Профиль с таким именем уже есть"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chart-profiles/{profile_id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Профиль",
        "operationId": "getChartProfile",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "profiles"
        ],
        "summary": "Переименовать или обновить конфигурацию",
        "operationId": "updateChartProfile",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          },
          "409": {
            "description": "Профиль с таким именем уже есть"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Удалить профиль",
        "operationId": "deleteChartProfile",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Не найдено"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chart-profiles/{profile_id}/use": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Запомнить как последний применённый",
        "operationId": "useChartProfile",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileConfig": {
        "properties": {
          "chart_timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Timeframe",
            "description": "TF графика при сохранении (необязательно)"
          },
          "indicators": {
            "items": {
              "$ref": "#/components/schemas/ProfileIndicator"
            },
            "type": "array",
            "title": "Indicators"
          },
          "layers": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Layers",
            "description": "Включённые слои (объём, ролл-маркеры, …)"
          },
          "style": {
            "additionalProperties": true,
            "type": "object",
            "title": "Style"
          }
        },
        "type": "object",
        "title": "ProfileConfig"
      },
      "ProfileIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id",
            "description": "Пусто — глобальный профиль, иначе для этого инструмента"
          },
          "config": {
            "$ref": "#/components/schemas/ProfileConfig"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProfileIn"
      },
      "ProfileIndicator": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Имя из `GET /indicators`"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "source_timeframe": {
            "type": "string",
            "title": "Source Timeframe",
            "description": "TF, по свечам которого считается"
          }
        },
        "type": "object",
        "required": [
          "name",
          "source_timeframe"
        ],
        "title": "ProfileIndicator"
      },
      "ProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "config": {
            "$ref": "#/components/schemas/ProfileConfig"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "instrument_id",
          "config",
          "created_at",
          "updated_at",
          "last_used_at"
        ],
        "title": "ProfileOut"
      },
      "ProfilePatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ProfilePatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}